// apps/scripts/src/seed/seedUtil/ProtoCourse.type.ts
import { z } from "zod";

export const protoCourseSchema = z.object({
  subject: z.string(),
  guessedSubject: z.string(),
  teacher: z.string(),
  room: z.string().optional(),
});

export type ProtoCourse = z.infer<typeof protoCourseSchema>;

export type ProtoCourseTime = {
  weekday: number;
  start: number;
  duration: number;
  weeks: "ODD" | "EVEN" | "BOTH";
};

export type ExtendedProtoCourse = ProtoCourse & {
  isChoosable: boolean;
  normalizedCourseId: string;
};

export type ProtoCourseWithTimes = ExtendedProtoCourse & {
  times: ProtoCourseTime[];
};